pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "permumark"
version = "0.1.0"
description = "Error Correction Enhanced Permutation Watermark for Transformers"
authors = [{ name = "Xiaokun Luan", email = "luanxiaokun@pku.edu.cn" }]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]
dependencies = [
    "datasets==2.21.0",
    "peft==0.12.0",
    "sagemath-standard==10.4",
    "scipy==1.14.1",
    "sympy==1.13.3",
    "tokenizers==0.19.1",
    "torch==2.4.0",
    "tqdm==4.67.1",
    "transformers==4.43.2"
]

[tool.isort]
profile = "black"
known_third_party = ["datasets"]
known_first_party = ["attacks", "eval_utils"]
